==> 02_Pandas/07_data_select_basic.ipynb <==
# 7. 데이터 선택
---

import pandas as pd
df = pd.read_csv('data/customers3.csv', index_col='고객ID')
df

# 복습할겸 한번 데이터 선택을 어떻게 했었는지 확인해봅시다.
df['거주도시']

df[['고객명','거주도시']]

### Column 선택

df.columns[0]

- 첫번째 컬럼 데이터 조회하기

# 활용
df[df.columns[0]] # df['고객명'] 과 동일한 동작

- 마지막 컬럼 데이터 조회하기

df[df.columns[-1]] # 재구매의사

### 슬라이싱
- 데이터프레임이나 시리즈에서 원하는 범위의 데이터(행 또는 열)를 잘라내는 기능
- 데이터를 특정 부분만 보거나 수정할 때 사용하며, 콜론(:)을 사용해 범위를 지정합니다.

> **파이썬 슬라이스 문법**
> - 기본사용법 `data[start:end:step]`
>   - start(시작위치) : 데이터를 자르기 시작할 인덱스
>   - end(끝 위치) : 데이터 자르기를 멈출 인덱스 (해당위치 데이터 포함X)
>   - step(간격) : 데이터를 가져올 간격 지정 (생략시 1칸씩 가져옴)

# 원본 확인해보기
df

df['고객명'][0:5] # 0~4번째 데이터

df[['고객명','나이']][:3] # 처음 3명의 이름, 키 정보를 가져옴

df[3:] # 3번째 행부터 끝까지 가져옴
