==> 02_Pandas/09_data_select_condition.ipynb <==
# 9. 조건에 따른 데이터 선택
---
- 데이터프레임에서 특정 조건(참/거짓)을 만족하는 행들만 골라내는 기능

import pandas as pd
df = pd.read_csv('data/customers3.csv', index_col='고객ID')
df

- 고객중 나이가 30세 이상인 고객

df['나이'] >=30

- True에 해당하는 데이터들만 조회 할 수도 있다.

filter = df['나이'] >=30
df[filter]

- 필터를 반전시킬 수 있다.

df[~filter]

df[filter]

loc와 응용해서 사용하기
- 보통 조건에 따라 원하는 데이터들만 보기위해 많이 사용한다.

# 나이가 30살 이상인 고객들의 고객명 데이터
filter = df['나이'] >= 30
df.loc[filter, '고객명']

df.loc[filter, ['고객명', '나이', '거주도시']]

Quiz
- 평균구매액(만원)이 100만원 이상인 사람들의 고객명과 나이를 조회해보자

filter = df['평균구매액(만원)'] >= 100
df.loc[filter, ['고객명', '나이']]

## 다양한 문법을 활용한 조건 설정

**& 엔드 조건**
- 평균구매액이 90만원 이하 이고 재구매의사(점)이이 85 이상인 고객 데이터(전부)조회하기

filter = (df['평균구매액(만원)'] <= 90) & (df['재구매의사(점)'] >= 85)
df.loc[filter]

**| or 조건**
- 나이가 25살 미만이거나 35이상인 고객 데이터를 조회하기

filter = (df['나이'] < 30) | (df['나이'] > 35)
df[filter]

**String 함수**

**startwith()**
- 문자열이 특정 글자로 시작하는지 확인하는 메서드 (특정글자로 시작하면 True 아니면 False)

filter = df['고객명'].str.startswith('박')
df[filter]

**isin()**
- 각 데이터가 특정 값 목록(리스트)에 포함되어 있는지 확인하는 메서드

interests = ['fashion']
#
filter =df['주요관심사'].str.lower().isin(interests)
df.loc[filter]

**contains()**
- 문자열이 특정 글자가 포함되는지 확인하는 메서드 (특정글자로 포함되면 True 아니면 False)

filter = df['고객명'].str.contains('태')
df[~filter]

- contains()을 사용할때 주의할점
    - 만약 데이터가 NaN으로 되어있다면 문제가 발생 할 수 있다.

filter = df['주요관심사'].str.contains('Sports', na=False)

filter

df.loc[filter]
