==> 02_Pandas/14_grouping.ipynb <==
# 14. 그룹화
---
- 동일한 값을 가진 것들끼리 합쳐서 통계 또는 평균 등의 값을 계산하기 위해 사용

import pandas as pd
df = pd.read_csv('data/customers3.csv', index_col='고객ID')
df

**`groupby()` 함수**
- 데이터를 특정 기준에 따라 그룹화하여 집계 연산을 수행한다.
- 이를 통해 데이터의 패턴을 발견하거나 요약 통계를 얻는 데 매우 유용하다.        
- 기본적인 그룹화`groupby()`는 주어진 열을 기준으로 데이터를 그룹화하고, <br>
    그 결과에 대해 다양한 집계 연산을 적용할 수 있다

group = df.grouby('주요관심사').get_group('Electronics')
group

df.grouby('거주도시').size()

# 주요관심사가 일렉틀닉스인 데이터의 재구매의사점수 평균
df.groupby('주요관심사').get_group('Electronics')['재구매의사(점)'].mean()

**`agg()`**
- 그룹화된 데이터에 대해 원하는 집계 연산을 적용하여 요약 통계를 얻을 수 있다.

데이터 집계함수를 사용해보기 위한 DataFrame 수정

# 첫번째 문자가 대문자인지 아닌지 대문자가 아니면 대문자로 변환

def capitalize(interest):
    # notnull() : NaN인지 아닌지 NaN 이면 False 아니면 True
    if pd.notnull(interest):
        return interest.capitalize()
    return 'Sports'


df['주요관심사'] = df['주요관심사'].apply(capitalize)
df

주요관심사별로 고객만족도와 재구매의사의 최소,최대,총합,평균 계산

aggregated = df.groupby('주요관심사').agg(
    {
        '재구매의사(점)' : ['min', 'max', 'sum','mean'],
        '고객만족도(점)' : ['min', 'max', 'sum','mean']
    }
)
aggregated
